ports antag town to yogstation
i hope at least this codebase won't have as many problems as the previous ones, although i can already notice that there's something wrong with Z levels and the spell system was reworked and is totally fucked. let's just hope it is not enough work for me to abandon this project the second time.